Add type annotations, Move to utils file, add \limits to screenshots
