+++ src/Settings.js
@@ -5,7 +5,7 @@
  * default settings.
  */
 
-import utils from "./utils";
+import utils from "./utils/utils";
 import ParseError from "./ParseError";
 import {Token} from "./Token";
