+++ src/functions/mclass.js
@@ -2,7 +2,7 @@
 import defineFunction, {ordargument} from "../defineFunction";
 import buildCommon from "../buildCommon";
 import mathMLTree from "../mathMLTree";
-import utils from "../utils";
+import utils from "../utils/utils";
 import type {AnyParseNode} from "../parseNode";
 
 import * as html from "../buildHTML";
